pyproject.toml
[project]
name = "exp1-wikitext-103"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.4",
    "tiktoken>=0.9.0",
    "torch>=2.6.0",
    "datasets>=2.18.0",
    "pytest>=8.0.0",
    "hydra-core>=1.3.2",
    "wandb>=0.19.9",
    "loguru>=0.7.3",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff]
line-length = 100
indent-width = 2

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
